[ upstream ] Return new code with settings back to the original `andark`
