--- apps/andark_w/ChangeLog
@@ -1,1 +1,0 @@
-0.01: Release based on andark v0.04
