--- apps/andark_w/README.md
@@ -1,17 +1,0 @@
-# Analog Clock with widgets
-
-## Features
-
-* second hand
-* date
-* battery percentage
-* swipeable widgets
-
-![logo](andark_w_screen.png)
-
-## Changes from the original Analog Clock (`andark`)
-
-* behaves correctly on the light theme (no change in the image)
-* date, battery and numbers are printed above hands
-* hour hand is slighly shorter to improve readability when minute hand is behind a number
-* widgets are loaded and are swipeable from the top
