--- apps/andark_w/app_icon.js
@@ -1,1 +1,0 @@
-require("heatshrink").decompress(atob("mEwxH+CiHXAhARMADoiHFRAzVJMQrMF7pseAH630RWw2Q67+EAAQnRMSYoFAYQwPNrAnGF9wuXP4xUHAY4tKZR4NEF54hRF6xNCJ4QFFF8bVXF6rmVF/avJGC7LNDqIaYF/4v/DUgvZDAQcTCYkHh0Og4FEhwMCAoovfEYQvpEYoECMwQ1GF9abCF8IkBAYQ2CBQgYEFyYvGKgYDDF4omSeCQoDGQgviXYiSERw4vjGxAvoS4guceBIvsMAaHDF8YwFAE4vvSA4v/SHgjCABgkkF74waX9zxuGYYw/GGAA=="))
